+++ esp-hal/src/soc/esp32s3/cpu_control.rs
@@ -172,6 +172,47 @@ unsafe fn internal_park_core(core: Cpu, park: bool) {
     }
 }
 
+/// Returns `true` if the specified core is currently running (not stalled).
+#[instability::unstable]
+pub fn is_running(core: Cpu) -> bool {
+    if core == Cpu::AppCpu {
+        // CORE_1_RUNSTALL in bit 0 -> needs to be 0 to not stall
+        // CORE_1_CLKGATE_EN in bit 1 -> needs to be 1 to even be enabled
+        let system = SYSTEM::regs();
+        let r = system.core_1_control_0().read();
+        if r.control_core_1_clkgate_en().bit_is_clear() || r.control_core_1_runstall().bit_is_set()
+        {
+            // If the core is not enabled we can take this shortcut
+            return false;
+        }
+    }
+
+    // sw_stall_appcpu_c1[5:0],  sw_stall_appcpu_c0[1:0]} == 0x86 will stall APP CPU
+    // sw_stall_procpu_c1[5:0],  reg_sw_stall_procpu_c0[1:0]} == 0x86 will stall PRO CPU
+    let is_stalled = match core {
+        Cpu::ProCpu => {
+            let c1 = LPWR::regs()
+                .sw_cpu_stall()
+                .read()
+                .sw_stall_procpu_c1()
+                .bits();
+            let c0 = LPWR::regs().options0().read().sw_stall_procpu_c0().bits();
+            (c1 << 2) | c0
+        }
+        Cpu::AppCpu => {
+            let c1 = LPWR::regs()
+                .sw_cpu_stall()
+                .read()
+                .sw_stall_appcpu_c1()
+                .bits();
+            let c0 = LPWR::regs().options0().read().sw_stall_appcpu_c0().bits();
+            (c1 << 2) | c0
+        }
+    };
+
+    is_stalled != 0x86
+}
+
 impl<'d> CpuControl<'d> {
     /// Creates a new instance of `CpuControl`.
     #[instability::unstable]